feat: Add new user API test scenario
Adds a new Cucumber feature file for testing the 'Get user details'
API endpoint on the 'https://reqres.in' sample API service. The test
verifies that the response has a status code of 200, the user ID
matches the expected value, and the email and first name are present
in the response.